requirements.txt
selenium==3.141.0
requests==2.25.1
python-dotenv==0.15.0